修复部分依赖问题，更新材质

+++ src/main/java/net/shelmarow/betterlockon/BetterLockOnClient.java
@@ -29,6 +29,6 @@ public class BetterLockOnClient {
 
     @SubscribeEvent
     static void onFovUpdate(ComputeFovModifierEvent event) {
-        event.setNewFovModifier((float) (event.getNewFovModifier() + BLOCameraSetting.fovOffset));
+        event.setNewFovModifier(event.getNewFovModifier() + BLOCameraSetting.fovOffset);
     }
 }
